+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0] - 2024-02-20
+
+### Added
+- Add filename of the model path as an alias
+- Add support for self extending the context window (SelfExtend)
+
 ## [1.10.0] - 2024-02-19
 
 ### Added